Add a base DeleteUploadView class
